adding component tests and using simple components
