+++ src/javascripts/components/navbar.js
@@ -1,0 +1,14 @@
+import React from 'react';
+import ShoppingCart from 'javascripts/components/shopping-cart';
+
+const Navbar = () => {
+  return (
+    <div className="navbar">
+      <div className="content navbar__content">
+        <ShoppingCart />
+      </div>
+    </div>
+  );
+};
+
+export default Navbar;
